requirements.txt
Django==5.2.8
djangorestframework==3.16.1  # <-- CORRECTED VERSION
channels==4.3.0
channels-redis==4.3.0
psycopg2-binary==2.9.7
gunicorn==21.2.0